CSHARP-3305: PR comments
